Maintenence<BootMe>: Restructured scripts directory and updated necessary changes
Spotify doesn't update bug

Fixed refresh buttons bug
