+++ scripts/backend/spoticry.py
@@ -6,7 +6,7 @@ from spotipy.oauth2 import SpotifyOAuth
 
 class Spoticry:
   def __init__(self):
-    base_dir = Path(__file__).resolve().parent.parent
+    base_dir = Path(__file__).resolve().parent.parent.parent
     env_path = base_dir/".env"
     load_dotenv(dotenv_path=env_path)
