+++ scripts/backend/trigger_actions.py
@@ -1,10 +1,10 @@
-import os 
 import webbrowser
 import subprocess
-import json
 from tkinter import messagebox
-from scripts.workflows import (kill_all_apps, clean_desktop, get_config)
-from scripts.spoticry import Spoticry
+from scripts.backend.workflows import (kill_all_apps, clean_desktop, get_config)
+from scripts.backend.spoticry import Spoticry
+
+import pdb
 
 def trigger_env(index):
   try: 
@@ -17,8 +17,11 @@ def trigger_env(index):
     for app in env["apps"]:
       subprocess.call(["open", app])
     
-    if env["spotify_type"] != "none":
+    # pdb.set_trace()
+
+    if env["spotify_url"] != "":
       spoticry = Spoticry()
+
       if env["spotify_type"] == "album":
         spoticry.start_album(env["spotify_id"])
       else:
